requirements.txt
cupy-cuda12x==13.6.0
matplotlib==3.10.6
numpy==2.3.3
pandas==2.3.2
Pillow==11.3.0
pygame==2.6.1
